libs/protoio: Move into internal
