+++ internal/protoio/io.go
