+++ internal/protoio/reader.go
