+++ internal/protoio/writer.go
