+++ p2p/conn/evil_secret_connection_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/cometbft/cometbft/crypto"
 	"github.com/cometbft/cometbft/crypto/ed25519"
 	cryptoenc "github.com/cometbft/cometbft/crypto/encoding"
-	"github.com/cometbft/cometbft/libs/protoio"
+	"github.com/cometbft/cometbft/internal/protoio"
 	tmp2p "github.com/cometbft/cometbft/proto/tendermint/p2p"
 )
